+++ modules/tinymce/src/core/demo/ts/demo/InlineDemo.ts
@@ -24,10 +24,10 @@ export default () => {
       { title: 'Some title 2', description: 'Some desc 2', content: '<div class="mceTmpl"><span class="cdate">cdate</span><span class="mdate">mdate</span>My content2</div>' }
     ],
     plugins: [
-      'autosave advlist autolink link image lists charmap print preview hr anchor pagebreak toc',
+      'autosave advlist autolink link image lists charmap preview hr anchor pagebreak toc',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste textpattern',
-      'codesample help noneditable print'
+      'codesample help noneditable'
     ]
   };