+++ modules/tinymce/src/core/demo/ts/demo/ResponsiveDemo.ts
@@ -67,7 +67,7 @@ export default () => {
     mobile: {
       theme: 'silver',
       plugins: [
-        'advlist autolink lists link image charmap print preview anchor',
+        'advlist autolink lists link image charmap preview anchor',
         'searchreplace visualblocks code fullscreen',
         'insertdatetime media table paste'
       ],
@@ -78,10 +78,10 @@ export default () => {
     },
     plugins: [
       'fullscreen help',
-      'autosave advlist autolink link image lists charmap print preview hr anchor pagebreak toc',
+      'autosave advlist autolink link image lists charmap preview hr anchor pagebreak toc',
       'searchreplace wordcount visualblocks visualchars code fullscreen insertdatetime media nonbreaking',
       'save table directionality emoticons template paste importcss textpattern',
-      'codesample help noneditable print'
+      'codesample help noneditable'
     ],
     // rtl_ui: true,
     add_unload_trigger: false,